Rename package security.viewers to security.users
