+++ server/test/unit/com/thoughtworks/go/config/security/GoConfigPipelinePermissionsAuthorityTest.java
@@ -19,8 +19,8 @@ package com.thoughtworks.go.config.security;
 import com.thoughtworks.go.config.Authorization;
 import com.thoughtworks.go.config.CruiseConfig;
 import com.thoughtworks.go.config.PipelineConfigs;
-import com.thoughtworks.go.config.security.viewers.AllowedViewers;
-import com.thoughtworks.go.config.security.viewers.Viewers;
+import com.thoughtworks.go.config.security.users.AllowedViewers;
+import com.thoughtworks.go.config.security.users.Viewers;
 import com.thoughtworks.go.helper.GoConfigMother;
 import com.thoughtworks.go.server.service.GoConfigService;
 import org.junit.After;
